feat: make loggable data available at handler level
