+++ src/controllers/source.ts
@@ -2,7 +2,7 @@ import { Context } from 'koa';
 import { MiscService } from '../services/misc';
 import { ServiceSelector } from '../helpers/serviceSelector';
 import { ControllerUtility } from './util';
-import logger from '../logger';
+import logger from '../logger/structured-logger';
 import { SourcePostTransformationService } from '../services/source/postTransformation';
 
 export class SourceController {
